# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Salmon API
  version: 0.15.1
  description: |
    Public HTTP API of the Salmon Wallet backend (`gol-salmon-api`).

    The API is organized as vertical slices per blockchain, mounted at
    `/v1/<chain>-{env}` (e.g. `/v1/solana-mainnet`), plus cross-chain
    surfaces (`/v1/networks`, `/v1/bridge`, `/v1/dapp`, CoinGecko market
    data, and the multichain balance endpoint) and unversioned service-info
    endpoints (`/health`, `/status`, `/ip`).

    Error payloads use the envelope `{ error: '<snake_case_code>',
    error_description: '<human message>' }` (see the `Error` schema), except
    the analytics ingest endpoint (`POST /v1/events`), which is a separate
    dependency-light Lambda and returns `{ error: '<code>' }` without
    `error_description`.

    Response shapes are defined by the `src/resources/` layer — that layer
    is the public contract. Shapes documented as `additionalProperties: true`
    are provider-enriched and carry extra fields beyond the stable ones
    listed here; the referenced resource file is the source of truth.
# The API is public and unauthenticated (CORS-restricted to *.salmonwallet.io
# origins, per-IP rate limited over /v1). Declared explicitly so tooling knows
# the absence of a security scheme is intentional.
security: []
servers:
  - url: https://api.salmonwallet.io
    description: production
  - url: http://localhost:3000/local
...
paths:
...
  /v1/exchange-rates:
    get:
      tags: [market-data]
      operationId: getExchangeRates
      summary: BTC-denominated fiat exchange rates
      description: Raw CoinGecko exchange-rates payload. Cached 900s.
      responses:
        '200':
          description: Exchange rates (raw CoinGecko shape)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/chart/{platform}/contract/{address}:
    get:
      tags: [market-data]
      operationId: getContractMarketChart
      summary: Market chart for a token by contract address (mint)
      description: |
        Charts any token CoinGecko lists for the platform, keyed by contract
        address instead of CoinGecko coin id. Same `days` normalization and
        long-term-fallback policy as `/v1/chart/{coinId}`.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
          description: CoinGecko asset-platform id (e.g. `solana`)
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Token contract address / mint
        - name: days
          in: query
          schema:
            type: string
          description: Numeric day count, `ytd`, or `max`
        - name: currency
          in: query
          schema:
            type: string
            default: usd
      responses:
        '200':
          description: Chart data
          content:
            application/json:
              schema:
                type: object
                required: [platform, contractAddress, currency, days, prices]
                properties:
                  platform:
                    type: string
                  contractAddress:
                    type: string
                  currency:
                    type: string
                  days: {}
                  prices:
                    type: array
                    items: { type: array, items: { type: number } }
                  marketCaps:
                    type: array
                    items: { type: array, items: { type: number } }
                  totalVolumes:
                    type: array
                    items: { type: array, items: { type: number } }
        '404':
          description: CoinGecko does not list this contract on this platform
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/chart/{coinId}:
    get:
      tags: [market-data]
      operationId: getMarketChart
      summary: Historical market chart data for a coin
      description: |
        Raw CoinGecko market-chart payload. `days` is normalized and a
        long-term-fallback policy is applied by the service (see the
        `coingecko-market-data` contract). Cached 300s.
      parameters:
        - name: coinId
          in: path
          required: true
          schema:
            type: string
          example: solana
        - name: days
          in: query
          required: false
          schema:
            type: string
          description: Number of days of history (normalized server-side)
        - name: currency
          in: query
          required: false
          schema:
            type: string
          example: usd
      responses:
        '200':
          description: Market chart data (raw CoinGecko shape, e.g. `prices`, `market_caps`, `total_volumes` arrays)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/coin/{coinId}:
    get:
      tags: [market-data]
      operationId: getCoinInfo
      summary: Coin detail
      description: Raw CoinGecko coin-detail payload. Cached 300s.
      parameters:
        - name: coinId
          in: path
          required: true
          schema:
            type: string
          example: bitcoin
        - name: currency
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Coin detail (raw CoinGecko shape)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/coin/{platform}/contract/{address}:
    get:
      tags: [market-data]
      operationId: getContractCoinInfo
      summary: Coin detail by token contract address
      description: Same mapped shape as `/v1/coin/{coinId}`, resolved by contract address (mint) for tokens without a known CoinGecko id. The response always includes the resolved CoinGecko `id`, so clients can cache it and switch to the coin-id paths. Cached 300s.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
          example: solana
        - name: address
          in: path
          required: true
          schema:
            type: string
          example: JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN
        - name: currency
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Coin detail (same shape as `/v1/coin/{coinId}`, `id` included)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Contract not listed on CoinGecko for the platform (`info_not_found`) — expected for the long tail of SPL tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  # ───────────────────────── dapp ─────────────────────────
...
components:
...
  responses:
    ServerError:
      description: Unexpected error (`server_error`), emitted by the final error middleware
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: |
        Standard error envelope. `error` is a stable snake_case code;
        `error_description` is a human-readable message. Codes in use:
        `bad_request`, `missing_parameter`, `invalid_parameter`,
        `too_many_mints`, `not_found` (catch-all 404), `nft_not_found`,
        `nfts_not_found`, `route_not_found`, `execution_failed`,
        `burn_not_supported`, `burn_transaction_too_large`,
        `burn_transaction`, `transfer_not_supported`,
        `transfer_transaction_too_large`, `transfer_transaction`,
        `server_error`.
      properties:
        error:
          type: string
          description: Stable snake_case error code
          example: bad_request
        error_description:
          type: string
          description: Human-readable message
      required: [error, error_description]
...